2013_12_15 - Updated and Finished?  Please see note below
The only thing I can't seem to get to look like the actual site is the
"Sign In" on the navigation bar.  I've got the links in place and the
color set to their version of blue.  After clicking, it should turn
green.  However, neither color visually looks right.  (They're OK if I
inspect element.)  Perhaps the font size, background color, or computer
(a Mac)?